refactor: encapsulte surface vm stuff in its vm
